feat: create simple demo day order Slack app
Created a scheduled AWS lambda function is scheduled
to run weekly every Thursday at 8am starting on 2021-01-07.
The function will check if the given Thursday is a
valid demo day. If it is, it will post a message and if not it will do
nothing.

There is also a simple command `/slash-get-list` endpoint that
is hooked up to the DemoDayOrder Slack app command `/get-demo-order`
so users can run `/get-demo-order` on the channel the app is added
to generate a list on demand. The endpoint is prefixed with 'slash' to
be explicit that this route is to be used as a slash command on Slack.

There are also endpoints scaffolded out to set up the app for interactive
Slack commands that I did not get to during the winter break. Ideally
we should be able to set a custom start time, verify scheduled dates
and delete a schedule. This will likely require setting up
a dynamoDB instance to save the start date and scheduled dates to verify against
instead of using the hard coded list of dates in the util function
`verify_is_valid_demo_day`